Move duplicated code in handlers to GesturesHandlerBase
