+++ Touch-Gestures-0.6.x/GesturesHandler.cs
@@ -1,75 +1,41 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
-using System.Numerics;
-using System.Threading;
 using OpenTabletDriver;
-using OpenTabletDriver.Desktop.Reflection;
 using OpenTabletDriver.Plugin;
 using OpenTabletDriver.Plugin.Attributes;
 using OpenTabletDriver.Plugin.DependencyInjection;
 using OpenTabletDriver.Plugin.Output;
 using OpenTabletDriver.Plugin.Platform.Pointer;
 using OpenTabletDriver.Plugin.Tablet;
-using OpenTabletDriver.Plugin.Tablet.Touch;
 using OTD.EnhancedOutputMode.Lib.Tools;
 using TouchGestures.Entities;
 using TouchGestures.Extensions;
-using TouchGestures.Lib.Entities;
-using TouchGestures.Lib.Entities.Gestures;
-using TouchGestures.Lib.Entities.Gestures.Bases;
-using TouchGestures.Lib.Entities.Tablet;
+using TouchGestures.Lib;
 
 namespace TouchGestures
 {
     [PluginName(PLUGIN_NAME)]
-    public class GesturesHandler : IPositionedPipelineElement<IDeviceReport>, IDisposable
+    public class GesturesHandler : GesturesHandlerBase, IPositionedPipelineElement<IDeviceReport>, IDisposable
     {
-        #region Constants
-
-        private const string PLUGIN_NAME = "Touch Gestures";
-
-        #endregion
-
         #region Fields
-        
+
         protected TouchSettings _touchSettings = TouchSettings.Default;
         protected IOutputMode? _outputMode;
-        protected GesturesDaemonBase? _daemon;
-        protected BindableProfile? _profile;
-        protected SharedTabletReference? _tablet;
         protected bool _awaitingDaemon;
-        private bool _hasPreviousGestureStarted;
 
         #endregion
 
         #region Initialization
 
-        public GesturesHandler()
+        public GesturesHandler() : base()
         {
-#if DEBUG
-            WaitForDebugger();
-            Log.Write(PLUGIN_NAME, "Debugger attached", LogLevel.Debug);
-#endif
-
             GesturesDaemonBase.DaemonLoaded += OnDaemonLoaded;
             _awaitingDaemon = true;
 
             BulletproofBindingBuilder.ChooseAsBuilder();
         }
 
-        private static void WaitForDebugger()
-        {
-            Console.WriteLine("Waiting for debugger to attach...");
-
-            while (!Debugger.IsAttached)
-            {
-                Thread.Sleep(100);
-            }
-        }
-
-        public virtual void Initialize()
+        public override void Initialize()
         {
             FetchTouchSettings();
 
@@ -90,7 +56,7 @@ namespace TouchGestures
                 Log.Write(PLUGIN_NAME, "Touch Gestures Daemon has not been enabled, please enable it in the 'Tools' tab", LogLevel.Error);
         }
 
-        protected virtual void InitializeCore(TabletReference tablet)
+        private void InitializeCore(TabletReference tablet)
         {
             _tablet = tablet.ToShared(_touchSettings);
 
@@ -113,6 +79,20 @@ namespace TouchGestures
 
         protected void FetchTouchSettings()
         {
+            if (_outputMode == null)
+                FetchOutputMode();
+
+            // then fetch the touch settings
+            var settings = _outputMode?.Elements.OfType<TouchSettings>().FirstOrDefault();
+
+            if (settings != null)
+                _touchSettings = settings;
+            else
+                Log.Write(PLUGIN_NAME, "Touch settings are null, using default values", LogLevel.Warning);
+        }
+
+        protected void FetchOutputMode()
+        {
             if (_Driver is Driver driver && Tablet != null)
             {
                 // fetch the device first
@@ -120,14 +100,6 @@ namespace TouchGestures
 
                 // then fetch the output mode
                 _outputMode = device?.OutputMode;
-
-                // then fetch the touch settings
-                var settings = _outputMode?.Elements.OfType<TouchSettings>().FirstOrDefault();
-
-                if (settings != null)
-                    _touchSettings = settings;
-                else
-                    Log.Write(PLUGIN_NAME, "Touch settings are null, using default values", LogLevel.Warning);
             }
         }
 
@@ -151,11 +123,6 @@ namespace TouchGestures
 
         #region Properties
 
-        public List<TapGesture> TapGestures { get; set; } = new();
-        public List<HoldGesture> HoldGestures { get; set; } = new();
-
-        public List<Gesture> NonConflictingGestures { get; set; } = new();
-
         [TabletReference]
         public TabletReference? Tablet { get; set; }
 
@@ -176,43 +143,8 @@ namespace TouchGestures
 
         public virtual void Consume(IDeviceReport report)
         {
-            if (report is ITouchReport touchReport)
-            {
-                if (_daemon != null && _daemon.IsReady && _touchSettings.IsTouchToggled)
-                {
-                    // Iterate through all conflicting gestures
-                    HandleConflictingGestures(TapGestures, touchReport);
-                    HandleConflictingGestures(HoldGestures, touchReport);
-
-                    // Iterate through all non-conflicting gestures
-                    foreach (var gesture in NonConflictingGestures)
-                        gesture.OnInput(touchReport.Touches);
-                }
-            }
-
+            Consume(report, _touchSettings.IsTouchToggled);
             Emit?.Invoke(report);
-        }
-
-        public void HandleConflictingGestures(IEnumerable<Gesture> gestures, ITouchReport touchReport)
-        {
-            _hasPreviousGestureStarted = false;
-
-            foreach (var gesture in gestures)
-            {
-                gesture.OnInput(touchReport.Touches);
-
-                // TODO: Ending it might not be the best move as simillar gesture might not work at all
-
-                // if the previous gesture has started, end any gesture after it
-                if (_hasPreviousGestureStarted)
-                {
-                    gesture.End();
-                    continue;
-                }
-
-                if (gesture.HasStarted)
-                    _hasPreviousGestureStarted = true;
-            }
         }
 
         #endregion
@@ -225,68 +157,16 @@ namespace TouchGestures
         public void OnDaemonLoaded(object? sender, EventArgs e)
             => Initialize();
 
-        public virtual void OnProfileChanged(object? sender, EventArgs e)
-        {
-            if (_profile == null)
-            {
-                Log.Write(PLUGIN_NAME, "Settings are null", LogLevel.Error);
-                return;
-            }
-
-            if (_tablet != null)
-            {
-                if (_tablet.TouchDigitizer != null && _tablet.TouchDigitizer.GetLPMM() != Vector2.Zero)
-                    _profile.UpdateLPMM(_tablet);
-                else
-                    Log.Write(PLUGIN_NAME, "LPMM is zero, this usually means that 'Touch Settings' hasn't been enabled or its maxes are set to zero", LogLevel.Error);
-            }
-
-            TapGestures.Clear();
-            HoldGestures.Clear();
-            NonConflictingGestures.Clear();
-
-            SortGestures();
-
-            // TODO : Non-Conflicting Tap & Hold gestures should be added to NonConflictingGestures
-            TapGestures.AddRange(_profile.TapGestures);
-            HoldGestures.AddRange(_profile.HoldGestures);
-
-            NonConflictingGestures.AddRange(_profile.SwipeGestures);
-            NonConflictingGestures.AddRange(_profile.PanGestures);
-            NonConflictingGestures.AddRange(_profile.PinchGestures);
-            NonConflictingGestures.AddRange(_profile.RotateGestures);
-
-            Log.Debug(PLUGIN_NAME, "Settings updated");
-        }
-
-        protected void SortGestures()
-        {
-            if (_profile == null)
-                return;
-
-            _profile.TapGestures.Sort((x, y) => x.RequiredTouchesCount > y.RequiredTouchesCount ? -1 : 1);
-            _profile.HoldGestures.Sort((x, y) => x.RequiredTouchesCount > y.RequiredTouchesCount ? -1 : 1);
-            _profile.SwipeGestures.Sort((x, y) => x.Direction > y.Direction ? -1 : 1);
-            _profile.PanGestures.Sort((x, y) => x.Direction > y.Direction ? -1 : 1);
-            _profile.PinchGestures.Sort((x, y) => x.IsInner ? -1 : 1);
-            _profile.RotateGestures.Sort((x, y) => x.IsClockwise ? -1 : 1);
-        }
-
         #endregion
 
         #region Interface Implementations
 
-        public void Dispose()
+        public override void Dispose()
         {
-            // Unsubscribe from events
-            if (_profile != null)
-                _profile.ProfileChanged -= OnProfileChanged;
+            base.Dispose();
 
             if (_awaitingDaemon)
                 GesturesDaemonBase.DaemonLoaded -= OnDaemonLoaded;
-
-            if (_tablet != null)
-                _daemon?.RemoveTablet(_tablet);
 
             _awaitingDaemon = false;
